+++ app/src/main/java/io/netbird/client/MyApplication.java
@@ -1,14 +1,9 @@
 package io.netbird.client;
 
 import android.app.Application;
-import android.content.IntentFilter;
 import android.content.SharedPreferences;
 
 import androidx.appcompat.app.AppCompatDelegate;
-import androidx.localbroadcastmanager.content.LocalBroadcastManager;
-
-import io.netbird.client.repository.VPNServiceRepository;
-import io.netbird.client.tool.NetworkChangeNotifier;
 
 public class MyApplication extends Application {
 
@@ -19,9 +14,5 @@ public class MyApplication extends Application {
         SharedPreferences prefs = getSharedPreferences("settings", MODE_PRIVATE);
         int themeMode = prefs.getInt("theme_mode", AppCompatDelegate.MODE_NIGHT_FOLLOW_SYSTEM);
         AppCompatDelegate.setDefaultNightMode(themeMode);
-    }
-
-    public VPNServiceRepository getVPNServiceRepository() {
-        return new VPNServiceRepository(this);
     }
 }
