--- app/src/main/java/io/netbird/client/repository/VPNServiceBindListener.java
@@ -1,5 +1,0 @@
-package io.netbird.client.repository;
-
-public interface VPNServiceBindListener {
-    void onServiceBind();
-}
